🐛 logs component is more stable
For the record, the previous logs component was printing lines on
multiple lines, and a change was made so a horizontal scrollbar
appeared on bottom so lines were on one line.

The computation was not removed and was causing a bug using the
vertical scrollbar or while scrolling into the view.

As part of this commit, the computation is removed, allowing a more
stable and a more performant component.

closes #88